Fix all CLang Warnings
This kills all the warnings emitted by Clang.  Mainly,
the remaining warnings are caused by mismatched signs.  Some
of these lead to true undefined behavior because:
  a)  Stat returns the file size as an off_t
  b)  Read takes an io_size of type size_t
  c)  Read returns the successful io_size as a ssize_t

So, in the long run, we need to check and make sure that we
never read or write buffers larger than SSIZE_MAX.  This
is pretty unlikely, but probably leads to problems if it
happens.

+++ src/base/target_pass.c
@@ -227,7 +227,7 @@ xdd_targetpass_task_setup(worker_data_t *wdp) {
 	}
 	 
 	// Figure out the transfer size to use for this I/O
-	if (tdp->td_current_bytes_remaining < tdp->td_xfer_size)
+	if (tdp->td_current_bytes_remaining < (uint64_t)tdp->td_xfer_size)
 		wdp->wd_task.task_xfer_size = tdp->td_current_bytes_remaining;
 	else wdp->wd_task.task_xfer_size = tdp->td_xfer_size;
